new module 'experimental' for experimental algorithms
