+++ postpic/helper.py
@@ -903,76 +903,3 @@ def kspace_propagate(kspace, dt, nsteps=1, **kwargs):
         return next(gen)
 
     return itertools.islice(gen, nsteps)
-
-
-def _kspace_propagate_adaptive_generator(field_in, axis=0,
-                                         yield_zeroth_step=False):
-    transform_state = field_in._transform_state()
-    if transform_state is None:
-        raise ValueError("kspace must have the same transform_state on all axes. "
-                         "Please make sure that either all axes 'live' in spatial domain or all "
-                         "axes 'live' in frequency domain.")
-
-    do_fft = transform_state
-
-    if do_fft:
-        complex_ex = field_in.fft()
-    else:
-        complex_ex = field_in
-
-    moving_window_vect = [0]*complex_ex.dimensions
-    moving_window_vect[axis] = 1
-
-    t = 0.0
-    x0 = np.mean(complex_ex.axes[axis].extent)
-
-    if yield_zeroth_step:
-        yield t, field_in
-
-    while True:
-        complex_ex = complex_ex.autoreduce()
-
-        # print(complex_ex.axes[0].physical_length, complex_ex.axes[1].physical_length)
-        lengths = [ax.physical_length for ax in complex_ex.axes]
-        transv_lengths = lengths[:]
-        del transv_lengths[axis]
-
-        distance = 0.5*np.mean(transv_lengths)
-        timestep = 1.0*distance / PhysicalConstants.c
-
-        # propagation distance since beginning
-        x = np.mean(complex_ex.axes[axis].extent) - x0 + distance
-
-        # box length in propagation direction
-        xlen = complex_ex.axes[axis].physical_length
-
-        # padding in prop. dir
-        xpad = distance / x * (0.5*xlen)
-
-        # pad distance in transverse directions and xpad in prop. dir
-        pad = [distance] * complex_ex.dimensions
-        pad[axis] = [xpad, 0.1*xpad]
-
-        # do padding and fft autopadding
-        complex_ex = complex_ex.pad(pad).fft_autopad()
-
-        # do propagation
-        complex_ex = kspace_propagate(complex_ex, timestep, moving_window_vect=moving_window_vect)
-        t += timestep
-
-        # remove low field strength outer region
-        complex_ex = complex_ex.autocutout(fractions=(0.01, 0.02))
-
-        if do_fft:
-            yield t, complex_ex.fft()
-        else:
-            yield t, complex_ex
-
-
-def kspace_propagate_adaptive(field_in, axis=0, t_final=None, **kwargs):
-    gen = _kspace_propagate_adaptive_generator(field_in, axis=axis, **kwargs)
-    while True:
-        t, f = next(gen)
-        yield t, f
-        if t_final and t > t_final:
-            break
